+++ package.json
@@ -14,13 +14,13 @@
     "scripts": {
         "clean": "rm -rf dist/ && rm -rf src/gen/declarations.ts src/gen/translate.ts souffle_parser_gen.js souffle_value_parser_gen.js",
         "transpile": "tsc",
-        "gen-translation-modules": "node scripts/gen_translation_modules.js && eslint --fix src/gen/translate.ts src/gen/ast_facts.ts",
+        "gen-translation-modules": "node scripts/gen_translation_modules.js && eslint --fix src/gen/translate.ts src/gen/ast_relations.ts",
         "copy-dl": "cp -r src/lib/analyses dist/lib && cp -r src/lib/detectors dist/lib && cp src/gen/ast.dl dist/gen",
         "copy-functors": "cp -r src/functors dist/",
         "build-functors": "./scripts/build_functors.sh",
         "build-preamble": "npm run gen-translation-modules",
         "compile-program": "ts-node scripts/compile_program.ts",
-        "build-full": "npm run clean && npm run build-preamble && npm run transpile && chmod a+x dist/bin/cli.js dist/bin/llm-features.js && npm run copy-dl && npm run build-functors && npm run copy-functors && npm run compile-program",
+        "full-build": "npm run clean && npm run build-preamble && npm run transpile && chmod a+x dist/bin/cli.js dist/bin/llm-features.js && npm run copy-dl && npm run build-functors && npm run copy-functors && npm run compile-program",
         "build": "npm run transpile && chmod a+x dist/bin/cli.js dist/bin/llm-features.js && npm run copy-dl",
         "quick-build": "npm run transpile && chmod a+x dist/bin/cli.js dist/bin/llm-features.js && npm run copy-dl",
         "test": "c8 mocha",
@@ -35,7 +35,7 @@
         "dotenv": "^16.4.5",
         "fs-extra": "^11.2.0",
         "solc-typed-ast": "^18.2.4",
-        "souffle.ts": "^0.3.0",
+        "souffle.ts": "^0.3.1",
         "sqlite": "^5.1.1",
         "sqlite3": "^5.1.7"
     },
