+++ scripts/compile_program.ts
@@ -1,3 +1,3 @@
-import { compileDatalog } from "../src/lib/souffle/souffle";
+import { compileDatalog } from "../src/lib/souffle";
 
 compileDatalog();
