+++ src/lib/souffle/index.ts
@@ -1,1 +1,2 @@
-export * from "./souffle";
+export * from "./compile";
+export * from "./analyze";
